reinstate useRunWorkflow
