+++ plugins/scaffolder-frontend-workflow/src/components/Form/Workflow.tsx
@@ -1,12 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { stringifyEntityRef } from '@backstage/catalog-model';
 import { Progress } from '@backstage/core-components';
 import { errorApiRef, useApi } from '@backstage/core-plugin-api';
-import {
-  scaffolderApiRef,
-  useCustomFieldExtensions,
-  useTemplateSecrets,
-} from '@backstage/plugin-scaffolder-react';
+import { useCustomFieldExtensions } from '@backstage/plugin-scaffolder-react';
 import {
   NextFieldExtensionOptions,
   WorkflowProps,
@@ -15,7 +11,7 @@ import {
 import { ReactNode, useCallback, useEffect } from 'react';
 import { FormWrapper } from './FormWrapper';
 import { JsonValue } from '@backstage/types';
-import { useAsync } from '@react-hookz/web';
+import { useRunWorkflow } from '../../hooks/useRunWorkflow';
 
 type Props = Pick<
   WorkflowProps,
@@ -30,11 +26,8 @@ export function Workflow({
   ...props
 }: Props): JSX.Element {
   const errorApi = useApi(errorApiRef);
-  const scaffolderApi = useApi(scaffolderApiRef);
-  const [taskId, setTaskId] = useState<string>();
   const customFieldExtensions =
     useCustomFieldExtensions<NextFieldExtensionOptions<any, any>>(children);
-  const { secrets } = useTemplateSecrets();
   const templateRef = stringifyEntityRef({
     kind: 'Template',
     namespace: namespace,
@@ -43,17 +36,7 @@ export function Workflow({
 
   const { loading, manifest, error } = useTemplateParameterSchema(templateRef);
 
-  const [state, { execute }] = useAsync(async function runScaffolderWorkflow({
-    values,
-  }) {
-    const { taskId: id } = await scaffolderApi.scaffold({
-      templateRef,
-      values,
-      secrets,
-    });
-
-    setTaskId(id);
-  });
+  const { state, execute, taskId } = useRunWorkflow({ templateRef });
 
   const handleNext = useCallback(
     async (formData: Record<string, JsonValue> | undefined) => {
@@ -72,7 +55,7 @@ export function Workflow({
   }, [error, errorApi]);
 
   // eslint-disable-next-line no-console
-  console.log({ state, taskId });
+  console.log({ state, taskId, execute });
 
   return (
     <>
