+++ plugins/scaffolder-frontend-workflow/src/hooks/useRunWorkflow.ts
@@ -1,0 +1,31 @@
+import { useApi } from '@backstage/core-plugin-api';
+import {
+  scaffolderApiRef,
+  useTemplateSecrets,
+} from '@backstage/plugin-scaffolder-react';
+import { useAsync } from '@react-hookz/web';
+import { useState } from 'react';
+
+interface Props {
+  templateRef: string;
+}
+
+export function useRunWorkflow({ templateRef }: Props) {
+  const scaffolderApi = useApi(scaffolderApiRef);
+  const [taskId, setTaskId] = useState<string>();
+  const { secrets } = useTemplateSecrets();
+
+  const [state, { execute }] = useAsync(async function runScaffolderWorkflow({
+    values,
+  }) {
+    const { taskId: id } = await scaffolderApi.scaffold({
+      templateRef,
+      values,
+      secrets,
+    });
+
+    setTaskId(id);
+  });
+
+  return { state, execute, taskId };
+}
